Revert "WIP Add sign-ups command + code refactoring"

+++ src/handlers.js
@@ -2,14 +2,13 @@ const R = require('ramda');
 const SlackTemplate = require('claudia-bot-builder').slackTemplate;
 const { InvalidCallbackActionError } = require('./handlers.errors');
 const { formatTimestamp, formatDescription, formatGameName, formatUser, formatMatch } = require('./formatting');
-const { chain, concurrent, validateUser, validateCallbackValue, parseCallbackValue } = require('./handlers.utils');
+const { parseCallbackValue } = require('./handlers.utils');
 
 const supportedCommands = [
     ['[tournaments]', 'list open tournaments'],
     ['whoami', 'show who you are on Challonge'],
     ['connect [<challonge_username>]', 'connect your Slack and Challonge accounts'],
     ['disconnect', 'disconnect your Slack and Challonge accounts'],
-    ['signup', 'sign up for a tournament'],
     ['next', 'list open matches in tournaments you are part of'],
     ['help', 'show this information'],
 ];
@@ -56,24 +55,24 @@ const listOpenTournaments = ({ challongeService, userRepository }) => async func
             .get();
 };
 
-const signUpUserCallback = chain(
-    validateCallbackValue('sign_up'),
-    concurrent(
-        validateUser,
-        ({ challongeService }) => async ({ callbackValue: tournamentId }) => {
-            const tournament = await challongeService.fetchTournament(tournamentId);
-            return tournament
-                ? Promise.resolve({ tournament })
-                : Promise.reject(new Error(`Tournament not found: ${tournamentId}`));
-        },
-    ),
-    ({ challongeService }) => async function ({ user, callbackValue: tournamentId, tournament }) {
-        await challongeService.addTournamentParticipant(tournamentId, user.challongeUsername);
-        return new SlackTemplate(`Awesome! You are now signed up for tournament *${tournament.name}.* :tada:`)
-            .replaceOriginal(false)
-            .get();
-    },
-);
+const signUpUserCallback = ({ challongeService, userRepository }) => async function ({ sender, originalRequest }) {
+    const user = await userRepository.getUser(sender);
+    if (!user) {
+        return unknownUserResponse;
+    }
+
+    const tournamentId = parseCallbackValue('sign_up', originalRequest);
+    if (!tournamentId) {
+        throw new InvalidCallbackActionError(originalRequest);
+    }
+
+    const tournamentPromise = challongeService.fetchTournament(tournamentId);
+    const addParticipantPromise = challongeService.addTournamentParticipant(tournamentId, user.challongeUsername);
+    const [tournament] = [await tournamentPromise, await addParticipantPromise];
+    return new SlackTemplate(`Awesome! You are now signed up for tournament *${tournament.name}.* :tada:`)
+        .replaceOriginal(false)
+        .get();
+};
 
 const showCurrentUser = ({ userRepository }) => async function ({ sender }) {
     const user = await userRepository.getUser(sender);
@@ -138,43 +137,14 @@ const logInUserCallback = ({ challongeService, userRepository }) => async functi
         .get();
 };
 
-const logOutUser = chain(
-    validateUser,
-    ({ userRepository }) => async function ({ sender }) {
-        await userRepository.deleteUser(sender);
-        return 'Okay, you are now forgotten. I hope to see you later! :wave:';
-    },
-);
-
-
-const signUpUser = chain(
-    validateUser,
-    ({ challongeService }) => async message => {
-        const { user } = message;
-        const notSignedUpTournaments = await challongeService.fetchTournamentsForMember(user.challongeEmailHash, false);
-        return notSignedUpTournaments.length > 0
-            ? Promise.resolve(notSignedUpTournaments)
-            : Promise.reject('There are currently no tournaments where you can sign up.');
-    },
-    () => async function (notSignedUpTournaments) {
-        const response = new SlackTemplate()
-            .addAttachment('signup')
-            .addText('What tournament do you want to join? :simple_smile:')
-            .addColor('#252830');
-        response.getLatestAttachment().actions = [
-            {
-                type: 'select',
-                text: 'Select...',
-                name: 'sign_up',
-                options: R.map(({ id, name }) => ({
-                    text: name,
-                    value: id,
-                }))(notSignedUpTournaments),
-            },
-        ];
-        return response.get();
-    },
-);
+const logOutUser = ({ userRepository }) => async function ({ sender }) {
+    const user = await userRepository.getUser(sender);
+    if (!user) {
+        return unknownUserResponse;
+    }
+    await userRepository.deleteUser(sender);
+    return 'Okay, you are now forgotten. I hope to see you later! :wave:';
+};
 
 const listNextMatches = ({ challongeService, userRepository }) => async function ({ sender }) {
     const user = await userRepository.getUser(sender);
@@ -216,10 +186,13 @@ const showUsage = () => function ({ originalRequest }) {
         .get();
 };
 
-const closeUsageCallback = chain(
-    validateCallbackValue('close'),
-    () => () => ({ delete_original: true }), // NOTE SlackTemplate doesn't support this flag
-);
+const closeUsageCallback = () => function ({ originalRequest }) {
+    const shouldClose = !!parseCallbackValue('close', originalRequest);
+    if (!shouldClose) {
+        throw new InvalidCallbackActionError(originalRequest);
+    }
+    return { delete_original: true }; // NOTE SlackTemplate doesn't support this flag
+};
 
 module.exports = {
     listOpenTournaments,
@@ -228,7 +201,6 @@ module.exports = {
     logInUser,
     logInUserCallback,
     logOutUser,
-    signUpUser,
     listNextMatches,
     showUsage,
     closeUsageCallback,
